Scaffold game and layout
Took the "machine" scene out of the UI layout with HBoxContainer as
it was messing up the physics of things.

See also: https://github.com/godotengine/godot/issues/74940

Workaround is to place the physics nodes hard-coded anchored to the
center of the game scene, and just "reserve" that space with a fixed
width middle control in the HBoxContainer used for the rest of the
game layout.