Change GUI and Sound setup
Yes, this commit comes out-of-place in a line of commits to get
drag and drop working. That's what happens on a hobby project.

Game layout needed tweaks anyways, and I wanted room to place some
buttons for audio.

Audio toggles and by-default-disabled music (at least in dev mode)
is super useful at this stage.