Tweak physics
Fixes #15

True, we could do much better than this. But we're close to the
release of the demo, and I don't want to risk big changes.

These tweaks make sure the main problem (that the ball sometimes
lies around on a bumper and not trigger on_exit_body) disappears
because of the bounciness.